relocate maincard folder and fix max value to inputs
