+++ src/app/router/Routes.tsx
@@ -6,7 +6,7 @@ import App from '@layout/app/App.component';
 import StreetcodeContent from '@streetcode/Streetcode.component';
 
 import AdminPage from '@/features/AdminPage/AdminPage.component';
-import Addnewmainblock from '@/features/AdminPage/MainBlock/AddNewMainblock';
+import Addnewmainblock from '@/features/AdminPage/NewStreetcode/MainBlock/AddNewMainblock';
 
 const router = createBrowserRouter(createRoutesFromElements(
     <Route path={FRONTEND_ROUTES.BASE} element={<App />}>
