+++ src/features/AdminPage/NewStreetcode/MainBlock/DragableTags/ClickableTagItem.tsx
