+++ src/features/AdminPage/NewStreetcode/MainBlock/MainBlockAdmin.style.scss
