+++ src/features/AdminPage/NewStreetcode/MainBlock/PreviewFileModal/PreviewFileModal.component.tsx
