+++ src/features/AdminPage/NewStreetcode/MainBlock/PreviewFileModal/PreviewFileModal.styles.scss
